scripts/pipeline/requirements.txt
# Lofi Pipeline Requirements
pretty_midi>=0.2.10
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
